backend/requirements.txt
Flask
Flask-Cors
gunicorn
SQLAlchemy
psycopg2-binary
python-dotenv
